+++ eth_defi/wstgbp/__init__.py
@@ -1,0 +1,1 @@
+"""Wren Staked tGBP tokenised asset support."""